Refactor server.rs
