+++ crates/key-server/src/tests/mod.rs
@@ -3,10 +3,12 @@
 
 use crate::externals::{add_package, add_upgraded_package};
 use crate::key_server_options::{KeyServerOptions, RetryConfig, RpcConfig, ServerMode};
+use crate::master_keys::MasterKeys;
 use crate::sui_rpc_client::SuiRpcClient;
 use crate::tests::KeyServerType::Open;
 use crate::types::Network;
-use crate::{from_mins, DefaultEncoding, MasterKeys, Server};
+use crate::utils::from_mins;
+use crate::{DefaultEncoding, Server};
 use crypto::ibe;
 use crypto::ibe::public_key_from_master_key;
 use fastcrypto::ed25519::Ed25519KeyPair;
