+++ crates/key-server/src/utils.rs
@@ -25,4 +25,48 @@ macro_rules! git_version {
         GIT_REVISION
     }};
 }
+
+use crate::types::IbeMasterKey;
+use anyhow::anyhow;
+use crypto::ibe::MASTER_KEY_LENGTH;
+use fastcrypto::encoding::Encoding;
+use fastcrypto::serde_helpers::ToFromByteArray;
 pub use git_version;
+use std::env;
+use std::time::Duration;
+use sui_types::base_types::ObjectID;
+
+/// Creates a [Duration] from a given number of minutes.
+/// Can be removed once the `Duration::from_mins` method is stabilized.
+pub const fn from_mins(mins: u16) -> Duration {
+    // safe cast since 64 bits is more than enough to hold 2^16 * 60 seconds
+    Duration::from_secs((mins * 60) as u64)
+}
+
+/// Read a byte array from an environment variable and decode it using the specified encoding.
+pub fn decode_byte_array<E: Encoding, const N: usize>(env_name: &str) -> anyhow::Result<[u8; N]> {
+    let hex_string =
+        env::var(env_name).map_err(|_| anyhow!("Environment variable {} must be set", env_name))?;
+    let bytes = E::decode(&hex_string)
+        .map_err(|_| anyhow!("Environment variable {} should be hex encoded", env_name))?;
+    bytes.try_into().map_err(|_| {
+        anyhow!(
+            "Invalid byte array length for environment variable {env_name}. Must be {N} bytes long"
+        )
+    })
+}
+
+/// Read a master key from an environment variable.
+pub fn decode_master_key<E: Encoding>(env_name: &str) -> anyhow::Result<IbeMasterKey> {
+    let bytes = decode_byte_array::<E, MASTER_KEY_LENGTH>(env_name)?;
+    IbeMasterKey::from_byte_array(&bytes)
+        .map_err(|_| anyhow!("Invalid master key for environment variable {env_name}"))
+}
+
+/// Read an ObjectID from an environment variable.
+pub fn decode_object_id(env_name: &str) -> anyhow::Result<ObjectID> {
+    let hex_string =
+        env::var(env_name).map_err(|_| anyhow!("Environment variable {} must be set", env_name))?;
+    ObjectID::from_hex_literal(&hex_string)
+        .map_err(|_| anyhow!("Invalid ObjectID for environment variable {env_name}"))
+}
